- fixed warning in QListView - Toolbar Handles are now drawin using QStyle, currently   it's a hack because if bin. compat. - With a click on a handle a toolbar can be hiddeen/shown like in   Netscape or KDE   now (and as Hidden is just another dock this can be also disabled   and enabled) - Updated mainwindow test for the new stuff
[git-p4: depot-paths = "//depot/qt/main/": change = 16085]

+++ src/widgets/qlistview.cpp
@@ -1089,7 +1089,7 @@ void QListViewItem::setText( int column, const QString &text )
 	return;
 
     l->text = text;
-    if ( column == lsc )
+    if ( column == (int)lsc )
 	lsc = Unsorted;
     QListView * lv = listView();
     int oldW = lv ? lv->columnWidth( column ) : 0;
